+++ framework/registerFn.php
@@ -8,7 +8,7 @@ function view(string $name, array $data = []): void
     if (count($data) > 0) {
         extract($data);
     }
-    require __DIR__ . '/../resources/views/' . str_replace('.', '/', $name) . '.php';
+    require Path::join(__DIR__,  '..', 'resources', 'views', str_replace('.', '/', $name) . '.php');
 }
 
 /** Render the given component */
@@ -17,7 +17,7 @@ function component(string $name, array $data = []): void
     if (count($data) > 0) {
         extract($data);
     }
-    require __DIR__ . '/../resources/views/components/' . str_replace('.', '/', $name) . '.php';
+    require Path::join(__DIR__, '..', 'resources', 'views', 'components', str_replace('.', '/', $name) . '.php');
 }
 
 /** Translate the given label into user language */