+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added Config class to make access to env settings easier
 - Added index.php that load the autoloader and the bootstrapping
 - Added bootstrapping to bioman CLI to get access to the global functions
+- Added facade Path with function for joining paths
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2